pyproject.toml
[project]
name = "test-agent"
version = "0.1.0"
description = "Simple LLM agent with tools"
authors = [
    { name = "Matvey Terentyev" }
]
requires-python = ">=3.11"

dependencies = [
    "httpx",
    "loguru",
    "python-dotenv>=1.2.2",
    "rich>=15.0.0",
]

[dependency-groups]
dev = [
    "pytest",
    "pytest-asyncio",
    "respx",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/agent_app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
markers = [
    "external: tests that require external LLM server",
]
